+++ lib/src/message.dart
@@ -2,6 +2,8 @@
 import 'dart:convert';
 import 'dart:typed_data';
 
+import 'client.dart';
+
 /// Message class
 class Message {
   ///subscriber id auto generate by client
@@ -9,13 +11,21 @@ class Message {
 
   /// subject  and replyto
   final String subject, replyTo;
+  final Client _client;
 
   ///payload of data in byte
-  final Uint8List payload;
+  final Uint8List data;
 
   ///constructure
-  Message(this.subject, this.sid, this.replyTo, this.payload);
+  Message(this.subject, this.sid, this.data, this._client, {this.replyTo});
 
   ///payload in string
-  String get payloadString => utf8.decode(payload);
+  String get string => utf8.decode(data);
+
+  ///Repond to message
+  bool respond(Uint8List data) {
+    if (replyTo == null || replyTo == '') return false;
+    _client.pub(replyTo, data);
+    return true;
+  }
 }